pyproject.toml
[tool.poetry]
name = "get-papers-list"
version = "0.1.0"
description = "Fetch research papers from PubMed API"
authors = ["You <your.email@example.com>"]
readme = "README.md"
packages = [{include = "get_papers_list", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
pandas = "^1.5.0"
biopython = "^1.79"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
black = "^22.8.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
get-papers-list = "get_papers_list.get_papers_list:main"
